Added endpoint_processing.py and README
