+++ lib/api_endpoint/oas3/api_openapi_type_handler.py
@@ -1,33 +1,6 @@
 import six
 from lib import utils
 
-def metamodel_to_swagger_type_converter(input_type):
-    """
-    Converts API Metamodel type to their equivalent Swagger type.
-    A tuple is returned. first value of tuple is main type.
-    second value of tuple has 'format' information, if available.
-    """
-    input_type = input_type.lower()
-    if input_type == 'date_time':
-        return 'string', 'date-time'
-    if input_type == 'secret':
-        return 'string', 'password'
-    if input_type == 'any_error':
-        return 'string', None
-    if input_type == 'dynamic_structure':
-        return 'object', None
-    if input_type == 'uri':
-        return 'string', 'uri'
-    if input_type == 'id':
-        return 'string', None
-    if input_type == 'long':
-        return 'integer', 'int64'
-    if input_type == 'double':
-        return 'number', 'double'
-    if input_type == 'binary':
-        return 'string', 'binary'
-    return input_type, None
-
 class typeHandler():
 
     def __init__(self):
@@ -59,7 +32,7 @@ class typeHandler():
 
 
     def visit_builtin(self, builtin_type, new_prop):
-        data_type, format_ = metamodel_to_swagger_type_converter(builtin_type)
+        data_type, format_ = utils.metamodel_to_swagger_type_converter(builtin_type)
         if 'type' in new_prop and new_prop['type'] == 'array':
             item_obj = {'type': data_type}
             new_prop['items'] = item_obj
@@ -69,7 +42,6 @@ class typeHandler():
             new_prop['type'] = data_type
             if format_ is not None:
                 new_prop['format'] = format_
-
 
     def visit_generic(self, generic_instantiation, new_prop, type_dict, structure_svc, enum_svc, ref_path, enable_filtering):
         if generic_instantiation.generic_type == 'OPTIONAL':
@@ -99,7 +71,7 @@ class typeHandler():
                 self.check_type(res_type, res_id, type_dict, structure_svc, enum_svc, ref_path, enable_filtering)
 
             elif generic_instantiation.map_value_type.category == 'BUILTIN':
-                new_type['additionalProperties'] = {'type': metamodel_to_swagger_type_converter(
+                new_type['additionalProperties'] = {'type': utils.metamodel_to_swagger_type_converter(
                     generic_instantiation.map_value_type.builtin_type)[0]}
                 
             elif generic_instantiation.map_value_type.category == 'GENERIC':
@@ -118,14 +90,6 @@ class typeHandler():
                 del new_prop['$ref']
 
 
-    def is_type_builtin(self, type_):
-        type_ = type_.lower()
-        typeset = {'binary', 'boolean', 'datetime', 'double', 'dynamicstructure', 'exception',
-                'id', 'long', 'opaque', 'secret', 'string', 'uri'}
-        if type_ in typeset:
-            return True
-        return False
-
     def get_structure_info(self, struct_type, structure_svc, enable_filtering):
         """
         Given a type, return its structure info, if the type is a structure.
@@ -140,10 +104,9 @@ class typeHandler():
                 structure_info.fields = [field for field in structure_info.fields if not utils.is_filtered(field.metadata, enable_filtering)]
                 return structure_info
         except Exception as ex:
-            eprint("Error fetching structure info for " + struct_type)
-            eprint(ex)
+            utils.eprint("Error fetching structure info for " + struct_type)
+            utils.eprint(ex)
             return None
-
 
     def process_structure_info(self, type_name, structure_info, type_dict, structure_svc, enum_svc, ref_path, enable_filtering):
         new_type = {'type': 'object', 'properties': {}}
@@ -191,7 +154,7 @@ class typeHandler():
         type_dict[type_name] = enum_type
 
     def check_type(self, resource_type, type_name, type_dict, structure_svc, enum_svc, ref_path, enable_filtering):
-        if type_name in type_dict or self.is_type_builtin(type_name):
+        if type_name in type_dict or utils.is_type_builtin(type_name):
             return
         if resource_type == 'com.vmware.vapi.structure':
             structure_info = self.get_structure_info(type_name, structure_svc, enable_filtering)
