+++ lib/api_endpoint/swagger2/api_metamodel2swagger.py
@@ -1,6 +1,6 @@
+from lib import utils
 from lib.api_endpoint.swagger2 import api_swagger_parameter_handler as api_swagg_ph
 from lib.api_endpoint.swagger2 import api_swagger_response_handler as api_swagg_rh
-TAG_SEPARATOR = '/'
 
 def get_path(operation_info, http_method, url, service_name, type_dict, structure_dict, enum_dict,
              operation_id, error_map, enable_filtering):
@@ -18,12 +18,14 @@ def get_path(operation_info, http_method, url, service_name, type_dict, structur
                                          errors,
                                          error_map, type_dict, structure_dict, enum_dict, service_name, operation_id, enable_filtering)
 
-    path = build_path(service_name,
+    path_obj = utils.build_path(service_name,
                       http_method,
                       url,
                       documentation, par_array, operation_id=operation_id,
                       responses=response_map,
                       consumes=consumes_json, produces=produces)
+    post_process_path(path_obj)
+    path = utils.add_basic_auth(path_obj)
     return path
 
 def find_consumes(method_type):
@@ -54,7 +56,7 @@ def process_put_post_patch_request(url, service_name, operation_name, params,
     todo: handle query, formData and header parameters
     """
     ## Path
-    path_param_list, other_param_list, new_url = api_swagg_ph.extract_path_parameters(params, url)
+    path_param_list, other_param_list, new_url = utils.extract_path_parameters(params, url)
     par_array = []
     for field_info in path_param_list:
         parx = api_swagg_ph.convert_field_info_to_swagger_parameter('path', field_info, type_dict,
@@ -62,7 +64,7 @@ def process_put_post_patch_request(url, service_name, operation_name, params,
         par_array.append(parx)
 
     ## Body
-    body_param_list, other_param_list =  api_swagg_ph.extract_body_parameters(other_param_list)
+    body_param_list, other_param_list =  utils.extract_body_parameters(other_param_list)
     
     if body_param_list:
         parx = api_swagg_ph.wrap_body_params(service_name, operation_name, body_param_list, type_dict,
@@ -71,7 +73,7 @@ def process_put_post_patch_request(url, service_name, operation_name, params,
             par_array.append(parx)
             
     ## Query 
-    query_param_list, other_param_list = api_swagg_ph.extract_query_parameters(other_param_list)
+    query_param_list, other_param_list = utils.extract_query_parameters(other_param_list)
     for query_param in query_param_list:
         parx = api_swagg_ph.convert_field_info_to_swagger_parameter('query', query_param, type_dict,
                                                         structure_svc, enum_svc, enable_filtering)
@@ -89,7 +91,7 @@ def process_put_post_patch_request(url, service_name, operation_name, params,
 
 def process_get_request(url, params, type_dict, structure_svc, enum_svc, enable_filtering):
     param_array = []
-    path_param_list, other_params_list, new_url = api_swagg_ph.extract_path_parameters(params, url)
+    path_param_list, other_params_list, new_url = utils.extract_path_parameters(params, url)
     
     for field_info in path_param_list:
         parameter_obj = api_swagg_ph.convert_field_info_to_swagger_parameter('path', field_info,
@@ -97,7 +99,7 @@ def process_get_request(url, params, type_dict, structure_svc, enum_svc, enable_
         param_array.append(parameter_obj)
 
     ## Query 
-    query_param_list, other_params_list = api_swagg_ph.extract_query_parameters(other_params_list)
+    query_param_list, other_params_list = utils.extract_query_parameters(other_params_list)
     for query_param in query_param_list:
         parameter_obj = api_swagg_ph.convert_field_info_to_swagger_parameter('query', query_param,
                                                             type_dict, structure_svc, enum_svc, enable_filtering)
@@ -113,7 +115,7 @@ def process_get_request(url, params, type_dict, structure_svc, enum_svc, enable_
     return param_array, new_url
 
 def process_delete_request(url, params, type_dict, structure_svc, enum_svc, enable_filtering):
-    path_param_list, other_params, new_url = api_swagg_ph.extract_path_parameters(params, url)
+    path_param_list, other_params, new_url = utils.extract_path_parameters(params, url)
     param_array = []
     for field_info in path_param_list:
         parx = api_swagg_ph.convert_field_info_to_swagger_parameter('path', field_info, type_dict,
@@ -124,50 +126,6 @@ def process_delete_request(url, params, type_dict, structure_svc, enum_svc, enab
                                                         structure_svc, enum_svc, enable_filtering)
         param_array.append(parx)
     return param_array, new_url
-
-def build_path(service_name, method, path, documentation, parameters, operation_id, responses, consumes,
-               produces):
-    """
-    builds swagger path object
-    :param service_name: name of service. ex com.vmware.vcenter.VM
-    :param method: type of method. ex put, post, get, patch
-    :param path: relative path to an individual endpoint
-    :param documentation: api documentation
-    :param parameters: input parameters for the api
-    :param responses: response of the api
-    :param consumes: expected media type format of api input
-    :param produces: expected media type format of api output
-    :return: swagger path object.
-    """
-    path_obj = {}
-    path_obj['tags'] = tags_from_service_name(service_name)
-    if method is not None:
-        path_obj['method'] = method
-    if path is not None:
-        path_obj['path'] = path
-    if documentation is not None:
-        path_obj['summary'] = documentation
-    if parameters is not None:
-        path_obj['parameters'] = parameters
-    if responses is not None:
-        path_obj['responses'] = responses
-    if consumes is not None:
-        path_obj['consumes'] = consumes
-    if produces is not None:
-        path_obj['produces'] = produces
-    if operation_id is not None:
-        path_obj['operationId'] = operation_id
-    post_process_path(path_obj)
-    add_basic_auth(path_obj)
-    return path_obj
-
-def tags_from_service_name(service_name):
-    """
-    Generates the tags based on the service name.
-    :param service_name: name of the service
-    :return: a list of tags
-    """
-    return [TAG_SEPARATOR.join(service_name.split('.')[3:])]
 
 def post_process_path(path_obj):
     # Temporary fixes necessary for generated spec files.
@@ -181,28 +139,4 @@ def post_process_path(path_obj):
 
     # Allow invoking $task operations from the api-explorer
     if path_obj['operationId'].endswith('$task'):
-        path_obj['path'] = add_query_param(path_obj['path'], 'vmw-task=true')
-
-def add_query_param(url, param):
-    """
-    Rudimentary support for adding a query parameter to a url.
-    Does nothing if the parameter is already there.
-    :param url: the input url
-    :param param: the parameter to add (in the form of key=value)
-    :return: url with added param, ?param or &param at the end
-    """
-    pre_param_symbol = '?'
-    query_index = url.find('?')
-    if query_index > -1:
-        if query_index == len(url):
-            pre_param_symbol = ''
-        elif url[query_index + 1:].find(param) > -1:
-            return url
-        else:
-            pre_param_symbol = '&'
-    return url + pre_param_symbol + param
-
-def add_basic_auth(path_obj):
-    """Add basic auth security requirement to paths requiring it."""
-    if path_obj['path'] == '/com/vmware/cis/session' and path_obj['method'] == 'post':
-        path_obj['security'] = [{'basic_auth': []}]
+        path_obj['path'] = utils.add_query_param(path_obj['path'], 'vmw-task=true')
